% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSNPs.R
\name{checkSNPs}
\alias{checkSNPs}
\title{Check and format SNP chromosome names}
\usage{
checkSNPs(snp_info, add_chr_prefix = TRUE)
}
\arguments{
\item{snp_info}{Data frame of genetic association results.
-required_columns A character vector of required column names. Default: c("CHR", "POS", "LP", "SNP").}

\item{add_chr_prefix}{Logical. Whether to add "chr" prefix to the "CHR" column if not already present. Default: FALSE.}
}
\value{
The modified data frame if all required columns are present. Otherwise, an error is thrown.
}
\description{
This function checks if the required columns are present in a given SNP data frame,
and optionally adds a "chr" prefix to the chromosome column if specified.
}
