% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulon.R
\name{createRegulon}
\alias{createRegulon}
\title{Construct TF-gene regulatory network (GRN)}
\usage{
createRegulon(
  single_cell,
  n_targets = 5,
  peak2gene_method = "Signac",
  infer_method = "glm"
)
}
\arguments{
\item{single_cell}{Input single-cell multiomics data (Seurat object)}

\item{n_targets}{Minimum number of targets required in a regulon (default = 5)}

\item{peak2gene_method}{Two methods to choose: 'Signac' and 'GREAT'}

\item{infer_method}{You can choose different inference methods: GLMs ('glm'),
regularized GLMs('glmnet','cv.glmnet'), or Bayesian regression models('brms')}
}
\value{
A list containing:
- grn: Filtered GRN with regulons meeting the gene threshold
- tf_names: List of TFs with sufficient regulons
}
\description{
Using single-cell multiomics data (scRNA-seq and scATAC-seq data) to construct a global TF-peak-gene regulatory network
}
