% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grn.R
\name{extract_grn}
\alias{extract_grn}
\title{Extract GRN Data}
\usage{
extract_grn(regulons, single_cell, infer_method = "glm")
}
\arguments{
\item{infer_method}{GRN inference method:
\itemize{
\item 'glm' (default): Generalized linear model
\item 'cv.glmnet': Regularized GLMs
\item 'brms': Bayesian regression models
\item 'bagging_ridge': Bagging ridge and Bayesian ridge (CellOracle)
\item 'xgb': XGBoost gradient-boosted Random Forest (SCENIC)
}}
}
\description{
Helper function to extract GRN data based on inference method.
}
