% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistances.R
\name{getDistances}
\alias{getDistances}
\title{Compute Euclidean Distances Between Target and Non-Target Scores}
\usage{
getDistances(target_results)
}
\arguments{
\item{target_results}{A data frame with three columns: RegulonName, Target, and NonTarget.}
}
\value{
A list containing:
- \code{between}: Pairwise distances between Target and NonTarget.
- \code{within_target}: Pairwise distances within the Target group.
- \code{within_nontarget}: Pairwise distances within the NonTarget group.
}
\description{
This function calculates the Euclidean distances between Target and NonTarget scores
from the input dataset, as well as the within distances for each group.
}
