% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnergyScore.R
\name{getEnergyScore}
\alias{getEnergyScore}
\title{Energy Score Calculation}
\usage{
getEnergyScore(regulon2disease_results, targetCelltype = 1)
}
\arguments{
\item{regulon2disease_results}{TRS results from regulon2disease() function.}

\item{targetCelltype}{Can be specified as either the index of the target cell type column (e.g., 1),
or the name of the target cell type (e.g., "Monocytes").}
}
\description{
Calculate the energy statistics of regulons in the target cell type compared to other cell types.
}
