% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleScore.R
\name{getModuleScore}
\alias{getModuleScore}
\title{Calculates the peak importance score for a given regulon}
\usage{
getModuleScore(eachModule)
}
\arguments{
\item{eachModule}{A single regulon; this is generated in the 'regulon2disease()' function.}
}
\value{
A data frame with two columns: \code{Target} and \code{Importance_weighted}.
The first row contains the TF name and its total importance score, followed by the targets with their highest importance scores.
}
\description{
It extracts the highest importance score for each target and calculates the total importance for the transcription factor (TF).
}
