% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakScore.R
\name{getPeakScore}
\alias{getPeakScore}
\title{Calculate the weighted importance score for each peak}
\usage{
getPeakScore(snp2peak_map, w1 = 0.3, w2 = 0.7)
}
\arguments{
\item{snp2peak_map}{Data frame generated by the 'snp2peak()' function.
Must contain columns 'snp_id', 'Significance', and 'Strength'.}

\item{w1}{Weight of peak strength (peak-gene correlation, e.g., -log10(p-value)). Default = 0.5.}

\item{w2}{Weight of peak significance (genetic relevance, e.g., -log10 SNP p-value). Default = 0.5.}
}
\value{
A data frame where missing values are replaced and an 'Importance_weighted' column is added.
}
\description{
Replace missing values (NA) and calculate weighted importance scores for peaks.
}
\details{
\itemize{
\item Replace NA in 'snp_id' with 'no_snp' and in 'Significance' with 0.
\item Calculate weighted importance scores for each peak using the formula: \code{w1 * Strength + w2 * Significance}.
}
}
