% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomWeight.R
\name{getRandomWeight}
\alias{getRandomWeight}
\title{Generate random weights for permutation}
\usage{
getRandomWeight(regulons, random_scores)
}
\arguments{
\item{regulons}{regulon matrix from 'snp2peak_map'}

\item{random_scores}{genes random Zscore and specificity score}
}
\value{
The function gives random importance weights for background genes
}
\description{
Generate random weights for permutation
}
