% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulonScore.R
\name{getRegulonScore}
\alias{getRegulonScore}
\title{Calculates the Trait-associated Regulon Score (TRS) for a given regulon}
\usage{
getRegulonScore(each_module_score, theta = 0.5, alpha = 1)
}
\arguments{
\item{each_module_score}{A single regulon matrix containing importance weights and specificity scores;
this is generated in the 'regulon2disease()' function.}

\item{theta}{Weight for integrating TF and gene scores (range: 0.1 to 1, default = 0.5).}

\item{alpha}{Flexibility in penalization (default = 1), used to scale the penalty term.}
}
\value{
A list containing:
- \code{SpecificityScore}: The computed M1 value (specificity score).
- \code{GeneRiskScore}: The computed M2 value (genetic risk score).
- \code{RegulonScore}: The final TRS score for the regulon.
}
\description{
This function combines the importance weight (genetic risk score) and specificity score
(cell type specificity) to calculate a TRS (Trait-associated Regulon Score) for a given regulon.
}
