% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRiskScore.R
\name{getRiskScore}
\alias{getRiskScore}
\title{Calculates the peak importance score for a given regulon}
\usage{
getRiskScore(eachModule, top_n = 5)
}
\arguments{
\item{eachModule}{A single regulon; this is generated in the 'regulon2disease()' function.}

\item{top_n}{top n targets of each TF to calculate the importance of the TF (default n=5)}
}
\value{
A data frame with two columns: \code{Target} and \code{Importance_weighted}.
The first row contains the TF name and its total importance score, followed by the targets with their highest importance scores.
}
\description{
It extracts the highest importance score for each target and calculates the total importance for the transcription factor (TF).
}
