% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPchr.R
\name{getSNPchr}
\alias{getSNPchr}
\title{Format SNP chromosome names}
\usage{
getSNPchr(snp_bed)
}
\arguments{
\item{snp_bed}{Data frame of genetic association results in .bed format}

\item{getSNPchr}{Logical, whether to add "chr" prefix to the chromosome column (default: TRUE)}
}
\value{
Updated data frame with "chr" prefix added to the chromosome column if missing
}
\description{
Format SNP chromosome names
}
