% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_scale.R
\name{max_min_scale}
\alias{max_min_scale}
\title{Max-min scaling}
\usage{
max_min_scale(x, epsilon = 1e-06)
}
\arguments{
\item{x}{Numeric vector to be scaled}

\item{epsilon}{Define a small epsilon to avoid zero division (default = 1e-6)}
}
\value{
Scaled vector with values ranging from 0 to 1, adjusted to avoid all-zero results
}
\description{
Max-min normalization with epsilon adjustment
}
