% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif2tf.R
\docType{data}
\name{motif2tf}
\alias{motif2tf}
\title{Motif to TF Mapping Dataset}
\format{
A data frame with 3 columns:
\describe{
\item{motif}{The name of the motif.}
\item{tf}{The transcription factor associated with the motif.}
\item{score}{The score associated with the motif.}
}
}
\usage{
motif2tf
}
\description{
A dataset mapping motifs to transcription factors with scores.
}
\examples{
data(motif2tf)
}
\keyword{datasets}
