% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\docType{data}
\name{motifs}
\alias{motifs}
\title{Motifs Dataset}
\format{
A data frame with columns:
\describe{
\item{motif}{The name of the motif.}
\item{score}{The score associated with the motif.}
}
}
\usage{
motifs
}
\description{
A dataset of custom motifs for analysis.
}
\examples{
data(motifs)
}
\keyword{datasets}
