% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation}
\alias{permutation}
\title{Monte Carlo permutation analysis}
\usage{
permutation(
  target_scores_background,
  tf_list_1,
  len_of_regulon,
  all_genes,
  top_genes = 500,
  theta = 0.5,
  pow = 1,
  mo = 1
)
}
\arguments{
\item{target_scores_background}{Data frame containing specificity scores of genes for each cell type}

\item{tf_list_1}{List of high-quality TFs excluding the target TF}

\item{len_of_regulon}{Number of genes and TFs in the regulon}

\item{all_genes}{Vector of all genes from phenotype associations (e.g., MAGMA results)}

\item{top_genes}{Number of top-ranked disease-specific genes for JSI enrichment (default = 500)}

\item{theta}{Weight to adjust the specificity score of target genes in each regulon
Range: 0.1 ~ 1 (default = 0.5)}

\item{pow}{Soft-power factor to emphasize strong specificity (default = 1)}

\item{mo}{Statistical model selection:
- mo = 1: Genetic weight model
- mo = 0: Non-genetic weight model}
}
\value{
Randomly computed specificity * JSI score for a regulon-disease link
}
\description{
Monte Carlo permutation for calculating random specificity * JSI scores
}
