% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phastConsElements20Mammals.R
\docType{data}
\name{phastConsElements20Mammals.UCSC.hg38}
\alias{phastConsElements20Mammals.UCSC.hg38}
\title{Conservation Elements for 20 Mammals (hg38)}
\format{
A GRanges object with the following columns:
\describe{
\item{seqnames}{The chromosome name.}
\item{start}{The start position of the region.}
\item{end}{The end position of the region.}
\item{score}{Conservation score.}
}
}
\source{
UCSC Genome Browser
}
\usage{
phastConsElements20Mammals.UCSC.hg38
}
\description{
A dataset of conservation elements for 20 mammals aligned to the hg38 human genome.
}
\keyword{datasets}
