% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegulon.R
\name{createRegulon}
\alias{createRegulon}
\title{Construct TF-gene regulatory network (GRN)}
\usage{
createRegulon(
  single_cell,
  n_targets = 5,
  peak2gene_method = "Signac",
  infer_method = "glm",
  tss_upstream = 1e+05,
  tss_downstream = 0,
  exclude_exon_regions = TRUE,
  conserved_regions = phastConsElements20Mammals.UCSC.hg38
)
}
\arguments{
\item{single_cell}{Input single-cell multiomics data (Seurat object)}

\item{n_targets}{Minimum number of targets required in a regulon (default = 5)}

\item{peak2gene_method}{Method for peak-to-gene linkage: 'Signac' or 'GREAT'}

\item{infer_method}{GRN inference method:
\itemize{
\item 'glm' (default): Generalized linear model
\item 'cv.glmnet': Regularized GLMs
\item 'brms': Bayesian regression models
\item 'bagging_ridge': Bagging ridge and Bayesian ridge (CellOracle)
\item 'xgb': XGBoost gradient-boosted Random Forest (SCENIC)
}}
}
\value{
A list containing:
\itemize{
\item grn: Filtered GRN with regulons meeting the gene threshold
\item tf_names: List of TFs with sufficient regulons
}
}
\description{
Use single-cell multiomics data (scRNA-seq and scATAC-seq) to construct a global TF-peak-gene regulatory network.
}
