% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistances.R
\name{getDistances}
\alias{getDistances}
\title{calcaulte Euclidean distance}
\usage{
getDistances(regulon_set1, regulon_set2)
}
\arguments{
\item{regulon_set1}{regulons' scMORE score in target cell type}

\item{regulon_set2}{regulons' scMORE score in other cell types}
}
\description{
Define a function to calculate pairwise distances
}
