% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnergyScore.R
\name{getEnergyScore}
\alias{getEnergyScore}
\title{Energy score calculation}
\usage{
getEnergyScore(regulon_set1, regulon_set2)
}
\arguments{
\item{regulon_set1}{regulons' scMORE score in target cell type}

\item{regulon_set2}{regulons' scMORE score in other cell types}
}
\description{
Calculate the energy statistics of regulons in target cell type compared to other cell types
}
