% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeakScore.R
\name{getPeakScore}
\alias{getPeakScore}
\title{Calculate the weighted importance score for each peak-gene}
\usage{
getPeakScore(snp2peak_map)
}
\arguments{
\item{snp2peak_map}{Data frame generated by the 'snp2peak()' function.
Must contain columns 'snp_id', 'Significance', 'Strength', and 'geneScores'.}
}
\value{
A data frame where missing values are replaced and an 'Importance_weighted' column is added.
}
\description{
Replace missing values (NA) and calculate weighted importance scores (GRS) for peaks.
}
\details{
\itemize{
\item Replace NA in 'snp_id' with 'no_snp'.
\item Replace NA in 'Significance', 'Strength', and 'geneScores' with their column mean.
\item Calculate importance scores for each peak using the formula: \code{Strength * Significance * geneScores}.
}
}
