% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecificity.R
\name{getSpecificity}
\alias{getSpecificity}
\title{get specificity in each cell type}
\usage{
getSpecificity(single_cell)
}
\arguments{
\item{single_cell}{Input single-cell data (Seurat object)}
}
\value{
A data frame with columns:
- genes: Gene or TF names
- scores: Specificity scores
- celltypes: Cell type names
}
\description{
Calculate specificity scores of TFs and target genes across all cell types
}
