% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak2gene.R
\name{peak2gene}
\alias{peak2gene}
\title{Link genes to genomic peaks}
\usage{
peak2gene(grn_outputs)
}
\arguments{
\item{grn_outputs}{A list containing GRN output data, with a 'Regions' column specifying genomic peak ranges.}
}
\value{
A data frame mapping genes to genomic peaks with associated log-transformed p-values (Strength).
}
\description{
Link genes to genomic peaks
}
