% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon2disease.R
\name{regulon2disease}
\alias{regulon2disease}
\title{Identify Cell Type-Specific Regulons Relevant to Disease}
\usage{
regulon2disease(
  grn_outputs,
  target_scores,
  geneRiskScores,
  snp_info,
  perm_n = 1000,
  theta = 0.5,
  alpha = 1,
  top_n = 5,
  buffer = 500,
  p1 = 0.05,
  p2 = 0.05,
  p3 = 0.05
)
}
\arguments{
\item{grn_outputs}{GRN (Gene Regulatory Network) outputs containing TF-gene relationships.}

\item{target_scores}{A matrix of genes and their TF specificity scores across cell types.
Columns include:
- \code{genes}: Gene names
- \code{scores}: TF specificity scores
- \code{celltypes}: Cell type annotations}

\item{geneRiskScores}{Gene-based genetic association results from MAGMA. Must include the following columns:
- \code{SYMBOL}: Gene symbols
- \code{logP}: Log-transformed p-values
- \code{ZSTAT}: Z-scores from MAGMA}

\item{snp_info}{SNP information used for the analysis.}

\item{perm_n}{Number of permutations for Monte Carlo simulation. Default: 1000.}

\item{theta}{Weighting factor for integrating TF and gene specificity scores. Range: 0.1 to 1. Default: 0.5.}

\item{alpha}{Flexibility parameter for penalization in the scoring model. Default: 1.}

\item{top_n}{Number of top targets for each TF used to calculate TF importance. Default: 5.}

\item{buffer}{Numeric value specifying the flanking region size for genomic peaks.
Extends the peak range upstream and downstream by the specified value.
For example, \code{buffer = 500} adds 500 bp on both sides of a peak. Default: 500 bp.}

\item{p1}{Threshold for statistical significance of the cell type-specificity score (CTS)
for each regulon. Default: 0.05.}

\item{p2}{Threshold for statistical significance of the genetic relevance score (GRS)
for each regulon. Default: 0.05.}

\item{p3}{Threshold for statistical significance of the trait-associated regulon score (TRS).
Default: 0.05.}
}
\value{
A data frame containing the following:
- \code{specificity}: Cell type-specificity scores for each regulon (CTS).
- \code{genetic_risk_score}: Genetic relevance scores for each regulon (GRS).
- \code{regulon_score}: Trait-associated regulon scores (TRS).
- \code{p_values}: Statistical significance values for specificity, risk, and regulon scores.
}
\description{
Function to identify cell type-specific regulons (TRS) that are associated with disease by integrating
cell type specificity scores(CTS) and genetic relevance scores (GRS).
}
