% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon2disease.R
\name{regulon2disease}
\alias{regulon2disease}
\title{Identify cell type-specific regulons relevant to disease}
\usage{
regulon2disease(
  grn_outputs,
  target_scores,
  geneRiskScores,
  snp_info,
  perm_n = 1000,
  theta = 0.5,
  alpha = 1,
  top_n = 5,
  buffer = 500
)
}
\arguments{
\item{grn_outputs}{GRN outputs containing TF-gene relationships}

\item{target_scores}{Matrix of genes and TF specificity scores across cell types (columns: genes, scores, celltypes)}

\item{geneRiskScores}{MAGMA-based gene association results (columns: SYMBOL, logP, ZSTAT)}

\item{snp_info}{Information about SNPs for the analysis}

\item{perm_n}{Number of permutations for Monte Carlo simulation (default = 1000)}

\item{theta}{Weight for integrating TF and gene scores (range: 0.1~1, default = 0.5)}

\item{alpha}{Flexibility in penalization (default = 1)}

\item{top_n}{Top n targets of each TF to calculate the importance of the TF (default = 5)}

\item{buffer}{Distance buffer (in base pairs) for SNP-to-peak mapping (default = 500bp), which means each peak will be extended by 500bp upstream and 500bp downstream.}
}
\value{
A data frame containing specificity, genetic risk score and regulon score and corresponding P values.
}
\description{
Identify cell type-specific regulons relevant to disease
}
