% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMore.R
\name{scMore}
\alias{scMore}
\title{Single-Cell MultiOmics Regulon Enrichment (scMORE)}
\usage{
scMore(
  single_cell,
  snp_info,
  gene_info,
  n_targets = 5,
  perm_n = 1000,
  theta = 0.5,
  alpha = 1,
  buffer = 500,
  top_n = 5,
  p1 = 0.05,
  p2 = 0.05,
  p3 = 0.05,
  method = "cosine",
  nSeed = 1234
)
}
\arguments{
\item{single_cell}{The input single-cell data, typically a Seurat or SingleCellExperiment object.}

\item{snp_info}{GWAS summary statistics. Must include the following columns:
\code{chr} (chromosome), \code{pos} (position), \code{rsID} (SNP ID), \code{beta} (effect size),
and \code{p-value} (significance level).}

\item{gene_info}{Gene-based genetic association results from MAGMA or FUMA.
The function supports results in standard MAGMA/FUMA formats.}

\item{n_targets}{Minimum number of target genes required in a given regulon. Default: 5.}

\item{perm_n}{Number of Monte Carlo permutations to perform for significance testing. Default: 1000.}

\item{theta}{Weighting factor to adjust the specificity score of target genes within each regulon.
Range: 0.1 to 1. Default: 0.5.}

\item{alpha}{Flexibility parameter for penalization in the inference model. Default: 1.}

\item{buffer}{Numeric value specifying the flanking region size for genomic peaks.
Extends the peak range upstream and downstream. For example, \code{buffer = 500}
extends 500 bp on both sides of a peak. Default: 500 bp.}

\item{top_n}{Number of top targets for each TF used to calculate its importance. Default: 5.}

\item{p1}{Threshold for statistical significance of the cell type-specificity score
for each regulon. Default: 0.05.}

\item{p2}{Threshold for statistical significance of the genetic risk score
for each regulon. Default: 0.05.}

\item{p3}{Threshold for statistical significance of the trait-associated regulon score (TRS).
Default: 0.05.}

\item{method}{Method used for calculating the cell type specificity of nodes (TFs or genes):
'cosine' (default) or 'average'.}

\item{nSeed}{Random seed for reproducibility (i.e., \code{set.seed()}). Default: 1234.}

\item{peak2gene_method}{Method for mapping peaks to genes. Options: \code{'Signac'} or \code{'GREAT'}.}

\item{infer_method}{Method for inference modeling. Options:
- Generalized Linear Models (\code{'glm'})
- Regularized GLMs (\code{'glmnet'}, \code{'cv.glmnet'})
- Bayesian regression models (\code{'brms'}).}
}
\value{
A list containing:
- Disease-relevant regulons
- Their genetic and cell-type associations
}
\description{
Main function for identifying disease-relevant regulons using single-cell data.
}
