% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMore.R
\name{scMore}
\alias{scMore}
\title{Single-Cell MultiOmics REgulon discovery (scMORE)}
\usage{
scMore(
  single_cell,
  snp_info,
  gene_info,
  n_targets = 5,
  top_genes = 500,
  perm_n = 1000,
  theta = 0.5,
  pow = 1,
  mo = 1,
  buffer = 50,
  nSeed = 1234
)
}
\arguments{
\item{single_cell}{The input single-cell data}

\item{snp_info}{GWAS summary statistics, must include columns: chr, pos, rsID, beta, p-value}

\item{gene_info}{Gene-based genetic association results based on MAGMA or FUMA
scMORE support the format of results from MAGMA and FUMA}

\item{n_targets}{Minimum number of targets required in a given regulon (default = 5)}

\item{top_genes}{Number of top-ranked disease-specific genes for JSI enrichment (default = 500)}

\item{perm_n}{Number of Monte Carlo permutations for significance testing (default = 1000)}

\item{theta}{Weight to adjust the specificity score of target genes in each regulon
Range: 0.1 ~ 1 (default = 0.5)}

\item{pow}{Soft-power factor to emphasize strong specificity (default = 1)}

\item{mo}{Statistical model selection:
- mo = 1: Genetic weight model
- mo = 0: Non-genetic weight model}

\item{buffer}{Numeric value specifying the flanking sizes (default = 0)
- extend peak range (upstream and downstream + 50bp)
- buffer = 50bp, represent add 50 bp to both upstream and downstream of peak}

\item{nSeed}{set the seed for random sampling, i.e., set.seed()}
}
\value{
A list containing disease-relevant regulons and their associations
}
\description{
Main function for identifying disease-relevant regulons using single-cell data
}
