% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp2peak.R
\name{snp2peak}
\alias{snp2peak}
\title{Link SNPs to genomic peaks}
\usage{
snp2peak(snp_info, peak2gene_strength, buffer = 500)
}
\arguments{
\item{snp_info}{Data frame containing SNP information (required columns: CHR, POS, LP, SNP)}

\item{peak2gene_strength}{List containing GRN output, with a 'Regions' column for peaks}

\item{buffer}{Numeric value specifying the flanking sizes (default = 500bp)
- extend peak range (both upstream and downstream + 500bp)
- buffer = 500bp, represent add 500 bp to both upstream and downstream of peak}
}
\value{
Data frame mapping SNPs to genomic peaks with associated log-transformed p-values
}
\description{
Link SNPs to genomic peaks
}
